Move make_date, to_char to datafusion-functions  (#9601)
* Move make_date, to_char to datafusion-functions

* Update post merge to remove datetime_expressions.rs

* Fix benchmarks.

* Cargo fmt.
