+++ datafusion/expr/src/expr_fn.rs
@@ -768,14 +768,6 @@ nary_scalar_expr!(
     "replace the substring of string that starts at the start'th character and extends for count characters with new substring"
 );
 
-// date functions
-scalar_expr!(
-    ToChar,
-    to_char,
-    datetime format,
-    "converts a date, time, timestamp or duration to a string based on the provided format"
-);
-scalar_expr!(MakeDate, make_date, year month day, "make a date from year, month and day component parts");
 scalar_expr!(Nanvl, nanvl, x y, "returns x if x is not NaN otherwise returns y");
 scalar_expr!(
     Iszero,
