+++ datafusion/functions/src/datetime/to_char.rs
@@ -15,123 +15,119 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! DateTime expressions
-
+use std::any::Any;
 use std::sync::Arc;
 
+use arrow::datatypes::DataType;
 use arrow::util::display::{ArrayFormatter, DurationFormat, FormatOptions};
-use arrow::{
-    array::{Array, ArrayRef, PrimitiveArray},
-    datatypes::DataType,
-};
-use arrow_array::builder::PrimitiveBuilder;
 use arrow_array::cast::AsArray;
-use arrow_array::types::{Date32Type, Int32Type};
-use arrow_array::StringArray;
-use chrono::prelude::*;
-use chrono::NaiveDate;
+use arrow_array::{Array, ArrayRef, StringArray};
+use arrow_schema::DataType::{Date32, Date64, Duration, Time32, Time64, Timestamp, Utf8};
+use arrow_schema::TimeUnit::{Microsecond, Millisecond, Nanosecond, Second};
 
 use datafusion_common::{exec_err, Result, ScalarValue};
-use datafusion_expr::ColumnarValue;
+use datafusion_expr::TypeSignature::Exact;
+use datafusion_expr::{
+    ColumnarValue, ScalarUDFImpl, Signature, Volatility, TIMEZONE_WILDCARD,
+};
 
-/// Returns a string representation of a date, time, timestamp or duration based
-/// on a Chrono pattern.
-///
-/// The syntax for the patterns can be found at
-/// <https://docs.rs/chrono/latest/chrono/format/strftime/index.html>
-///
-/// # Examples
-///
-/// ```ignore
-/// # use chrono::prelude::*;
-/// # use datafusion::prelude::*;
-/// # use datafusion::error::Result;
-/// # use datafusion_common::ScalarValue::TimestampNanosecond;
-/// # use std::sync::Arc;
-/// # use arrow_array::{Date32Array, RecordBatch, StringArray};
-/// # use arrow_schema::{DataType, Field, Schema};
-/// # #[tokio::main]
-/// # async fn main() -> Result<()> {
-/// let schema = Arc::new(Schema::new(vec![
-///     Field::new("values", DataType::Date32, false),
-///     Field::new("patterns", DataType::Utf8, false),
-/// ]));
-///
-/// let batch = RecordBatch::try_new(
-///     schema,
-///     vec![
-///         Arc::new(Date32Array::from(vec![
-///             18506,
-///             18507,
-///             18508,
-///             18509,
-///         ])),
-///         Arc::new(StringArray::from(vec![
-///             "%Y-%m-%d",
-///             "%Y:%m:%d",
-///             "%Y%m%d",
-///             "%d-%m-%Y",
-///         ])),
-///     ],
-/// )?;
-///
-/// let ctx = SessionContext::new();
-/// ctx.register_batch("t", batch)?;
-/// let df = ctx.table("t").await?;
-///
-/// // use the to_char function to convert col 'values',
-/// // to strings using patterns in col 'patterns'
-/// let df = df.with_column(
-///     "date_str",
-///     to_char(col("values"), col("patterns"))
-/// )?;
-/// // Note that providing a scalar value for the pattern
-/// // is more performant
-/// let df = df.with_column(
-///     "date_str2",
-///     to_char(col("values"), lit("%d-%m-%Y"))
-/// )?;
-/// // literals can be used as well with dataframe calls
-/// let timestamp = "2026-07-08T09:10:11"
-///     .parse::<NaiveDateTime>()
-///     .unwrap()
-///     .with_nanosecond(56789)
-///     .unwrap()
-///     .timestamp_nanos_opt()
-///     .unwrap();
-/// let df = df.with_column(
-///     "timestamp_str",
-///     to_char(lit(TimestampNanosecond(Some(timestamp), None)), lit("%d-%m-%Y %H:%M:%S"))
-/// )?;
-///
-/// df.show().await?;
-///
-/// # Ok(())
-/// # }
-/// ```
-pub fn to_char(args: &[ColumnarValue]) -> Result<ColumnarValue> {
-    if args.len() != 2 {
-        return exec_err!("to_char function requires 2 arguments, got {}", args.len());
+#[derive(Debug)]
+pub(super) struct ToCharFunc {
+    signature: Signature,
+    aliases: Vec<String>,
+}
+
+impl ToCharFunc {
+    pub fn new() -> Self {
+        Self {
+            signature: Signature::one_of(
+                vec![
+                    Exact(vec![Date32, Utf8]),
+                    Exact(vec![Date64, Utf8]),
+                    Exact(vec![Time32(Millisecond), Utf8]),
+                    Exact(vec![Time32(Second), Utf8]),
+                    Exact(vec![Time64(Microsecond), Utf8]),
+                    Exact(vec![Time64(Nanosecond), Utf8]),
+                    Exact(vec![Timestamp(Second, None), Utf8]),
+                    Exact(vec![
+                        Timestamp(Second, Some(TIMEZONE_WILDCARD.into())),
+                        Utf8,
+                    ]),
+                    Exact(vec![Timestamp(Millisecond, None), Utf8]),
+                    Exact(vec![
+                        Timestamp(Millisecond, Some(TIMEZONE_WILDCARD.into())),
+                        Utf8,
+                    ]),
+                    Exact(vec![Timestamp(Microsecond, None), Utf8]),
+                    Exact(vec![
+                        Timestamp(Microsecond, Some(TIMEZONE_WILDCARD.into())),
+                        Utf8,
+                    ]),
+                    Exact(vec![Timestamp(Nanosecond, None), Utf8]),
+                    Exact(vec![
+                        Timestamp(Nanosecond, Some(TIMEZONE_WILDCARD.into())),
+                        Utf8,
+                    ]),
+                    Exact(vec![Duration(Second), Utf8]),
+                    Exact(vec![Duration(Millisecond), Utf8]),
+                    Exact(vec![Duration(Microsecond), Utf8]),
+                    Exact(vec![Duration(Nanosecond), Utf8]),
+                ],
+                Volatility::Immutable,
+            ),
+            aliases: vec![String::from("date_format")],
+        }
+    }
+}
+
+impl ScalarUDFImpl for ToCharFunc {
+    fn as_any(&self) -> &dyn Any {
+        self
     }
 
-    match &args[1] {
-        // null format, use default formats
-        ColumnarValue::Scalar(ScalarValue::Utf8(None))
-        | ColumnarValue::Scalar(ScalarValue::Null) => {
-            _to_char_scalar(args[0].clone(), None)
+    fn name(&self) -> &str {
+        "to_char"
+    }
+
+    fn signature(&self) -> &Signature {
+        &self.signature
+    }
+
+    fn return_type(&self, _arg_types: &[DataType]) -> Result<DataType> {
+        Ok(Utf8)
+    }
+
+    fn invoke(&self, args: &[ColumnarValue]) -> Result<ColumnarValue> {
+        if args.len() != 2 {
+            return exec_err!(
+                "to_char function requires 2 arguments, got {}",
+                args.len()
+            );
         }
-        // constant format
-        ColumnarValue::Scalar(ScalarValue::Utf8(Some(format))) => {
-            // invoke to_char_scalar with the known string, without converting to array
-            _to_char_scalar(args[0].clone(), Some(format))
+
+        match &args[1] {
+            // null format, use default formats
+            ColumnarValue::Scalar(ScalarValue::Utf8(None))
+            | ColumnarValue::Scalar(ScalarValue::Null) => {
+                _to_char_scalar(args[0].clone(), None)
+            }
+            // constant format
+            ColumnarValue::Scalar(ScalarValue::Utf8(Some(format))) => {
+                // invoke to_char_scalar with the known string, without converting to array
+                _to_char_scalar(args[0].clone(), Some(format))
+            }
+            ColumnarValue::Array(_) => _to_char_array(args),
+            _ => {
+                exec_err!(
+                    "Format for `to_char` must be non-null Utf8, received {:?}",
+                    args[1].data_type()
+                )
+            }
         }
-        ColumnarValue::Array(_) => _to_char_array(args),
-        _ => {
-            exec_err!(
-                "Format for `to_char` must be non-null Utf8, received {:?}",
-                args[1].data_type()
-            )
-        }
+    }
+
+    fn aliases(&self) -> &[String] {
+        &self.aliases
     }
 }
 
@@ -143,14 +139,14 @@ fn _build_format_options<'a>(
         return Ok(FormatOptions::new());
     };
     let format_options = match data_type {
-        DataType::Date32 => FormatOptions::new().with_date_format(Some(format)),
-        DataType::Date64 => FormatOptions::new().with_datetime_format(Some(format)),
-        DataType::Time32(_) => FormatOptions::new().with_time_format(Some(format)),
-        DataType::Time64(_) => FormatOptions::new().with_time_format(Some(format)),
-        DataType::Timestamp(_, _) => FormatOptions::new()
+        Date32 => FormatOptions::new().with_date_format(Some(format)),
+        Date64 => FormatOptions::new().with_datetime_format(Some(format)),
+        Time32(_) => FormatOptions::new().with_time_format(Some(format)),
+        Time64(_) => FormatOptions::new().with_time_format(Some(format)),
+        Timestamp(_, _) => FormatOptions::new()
             .with_timestamp_format(Some(format))
             .with_timestamp_tz_format(Some(format)),
-        DataType::Duration(_) => FormatOptions::new().with_duration_format(
+        Duration(_) => FormatOptions::new().with_duration_format(
             if "ISO8601".eq_ignore_ascii_case(format) {
                 DurationFormat::ISO8601
             } else {
@@ -237,243 +233,19 @@ fn _to_char_array(args: &[ColumnarValue]) -> Result<ColumnarValue> {
     }
 }
 
-/// make_date(year, month, day) SQL function implementation
-pub fn make_date(args: &[ColumnarValue]) -> Result<ColumnarValue> {
-    if args.len() != 3 {
-        return exec_err!(
-            "make_date function requires 3 arguments, got {}",
-            args.len()
-        );
-    }
-
-    // first, identify if any of the arguments is an Array. If yes, store its `len`,
-    // as any scalar will need to be converted to an array of len `len`.
-    let len = args
-        .iter()
-        .fold(Option::<usize>::None, |acc, arg| match arg {
-            ColumnarValue::Scalar(_) => acc,
-            ColumnarValue::Array(a) => Some(a.len()),
-        });
-
-    let is_scalar = len.is_none();
-    let array_size = if is_scalar { 1 } else { len.unwrap() };
-
-    let years = args[0].cast_to(&DataType::Int32, None)?;
-    let months = args[1].cast_to(&DataType::Int32, None)?;
-    let days = args[2].cast_to(&DataType::Int32, None)?;
-
-    // since the epoch for the date32 datatype is the unix epoch
-    // we need to subtract the unix epoch from the current date
-    // note this can result in a negative value
-    let unix_days_from_ce = NaiveDate::from_ymd_opt(1970, 1, 1)
-        .unwrap()
-        .num_days_from_ce();
-
-    let mut builder: PrimitiveBuilder<Date32Type> = PrimitiveArray::builder(array_size);
-
-    let construct_date_fn = |builder: &mut PrimitiveBuilder<Date32Type>,
-                             year: i32,
-                             month: i32,
-                             day: i32,
-                             unix_days_from_ce: i32|
-     -> Result<()> {
-        let Ok(m) = u32::try_from(month) else {
-            return exec_err!("Month value '{month:?}' is out of range");
-        };
-        let Ok(d) = u32::try_from(day) else {
-            return exec_err!("Day value '{day:?}' is out of range");
-        };
-
-        let date = NaiveDate::from_ymd_opt(year, m, d);
-
-        match date {
-            Some(d) => builder.append_value(d.num_days_from_ce() - unix_days_from_ce),
-            None => return exec_err!("Unable to parse date from {year}, {month}, {day}"),
-        };
-        Ok(())
-    };
-
-    let scalar_value_fn = |col: &ColumnarValue| -> Result<i32> {
-        let ColumnarValue::Scalar(s) = col else {
-            return exec_err!("Expected scalar value");
-        };
-        let ScalarValue::Int32(Some(i)) = s else {
-            return exec_err!("Unable to parse date from null/empty value");
-        };
-        Ok(*i)
-    };
-
-    // For scalar only columns the operation is faster without using the PrimitiveArray
-    if is_scalar {
-        construct_date_fn(
-            &mut builder,
-            scalar_value_fn(&years)?,
-            scalar_value_fn(&months)?,
-            scalar_value_fn(&days)?,
-            unix_days_from_ce,
-        )?;
-    } else {
-        let to_primitive_array = |col: &ColumnarValue,
-                                  scalar_count: usize|
-         -> Result<PrimitiveArray<Int32Type>> {
-            match col {
-                ColumnarValue::Array(a) => Ok(a.as_primitive::<Int32Type>().to_owned()),
-                _ => {
-                    let v = scalar_value_fn(col).unwrap();
-                    Ok(PrimitiveArray::<Int32Type>::from_value(v, scalar_count))
-                }
-            }
-        };
-
-        let years = to_primitive_array(&years, array_size).unwrap();
-        let months = to_primitive_array(&months, array_size).unwrap();
-        let days = to_primitive_array(&days, array_size).unwrap();
-        for i in 0..array_size {
-            construct_date_fn(
-                &mut builder,
-                years.value(i),
-                months.value(i),
-                days.value(i),
-                unix_days_from_ce,
-            )?;
-        }
-    }
-
-    let arr = builder.finish();
-
-    if is_scalar {
-        // If all inputs are scalar, keeps output as scalar
-        Ok(ColumnarValue::Scalar(ScalarValue::Date32(Some(
-            arr.value(0),
-        ))))
-    } else {
-        Ok(ColumnarValue::Array(Arc::new(arr)))
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::sync::Arc;
-
-    use arrow::array::{ArrayRef, Int64Array};
+    use crate::datetime::to_char::ToCharFunc;
     use arrow_array::{
-        Date32Array, Date64Array, Int32Array, Time32MillisecondArray, Time32SecondArray,
-        Time64MicrosecondArray, Time64NanosecondArray, TimestampMicrosecondArray,
-        TimestampMillisecondArray, TimestampNanosecondArray, TimestampSecondArray,
-        UInt32Array,
+        Array, ArrayRef, Date32Array, Date64Array, StringArray, Time32MillisecondArray,
+        Time32SecondArray, Time64MicrosecondArray, Time64NanosecondArray,
+        TimestampMicrosecondArray, TimestampMillisecondArray, TimestampNanosecondArray,
+        TimestampSecondArray,
     };
-
+    use chrono::{NaiveDateTime, Timelike};
     use datafusion_common::ScalarValue;
-
-    use super::*;
-
-    #[test]
-    fn test_make_date() {
-        let res = make_date(&[
-            ColumnarValue::Scalar(ScalarValue::Int32(Some(2024))),
-            ColumnarValue::Scalar(ScalarValue::Int64(Some(1))),
-            ColumnarValue::Scalar(ScalarValue::UInt32(Some(14))),
-        ])
-        .expect("that make_date parsed values without error");
-
-        if let ColumnarValue::Scalar(ScalarValue::Date32(date)) = res {
-            assert_eq!(19736, date.unwrap());
-        } else {
-            panic!("Expected a scalar value")
-        }
-
-        let res = make_date(&[
-            ColumnarValue::Scalar(ScalarValue::Int64(Some(2024))),
-            ColumnarValue::Scalar(ScalarValue::UInt64(Some(1))),
-            ColumnarValue::Scalar(ScalarValue::UInt32(Some(14))),
-        ])
-        .expect("that make_date parsed values without error");
-
-        if let ColumnarValue::Scalar(ScalarValue::Date32(date)) = res {
-            assert_eq!(19736, date.unwrap());
-        } else {
-            panic!("Expected a scalar value")
-        }
-
-        let res = make_date(&[
-            ColumnarValue::Scalar(ScalarValue::Utf8(Some("2024".to_string()))),
-            ColumnarValue::Scalar(ScalarValue::LargeUtf8(Some("1".to_string()))),
-            ColumnarValue::Scalar(ScalarValue::Utf8(Some("14".to_string()))),
-        ])
-        .expect("that make_date parsed values without error");
-
-        if let ColumnarValue::Scalar(ScalarValue::Date32(date)) = res {
-            assert_eq!(19736, date.unwrap());
-        } else {
-            panic!("Expected a scalar value")
-        }
-
-        let years = Arc::new((2021..2025).map(Some).collect::<Int64Array>());
-        let months = Arc::new((1..5).map(Some).collect::<Int32Array>());
-        let days = Arc::new((11..15).map(Some).collect::<UInt32Array>());
-        let res = make_date(&[
-            ColumnarValue::Array(years),
-            ColumnarValue::Array(months),
-            ColumnarValue::Array(days),
-        ])
-        .expect("that make_date parsed values without error");
-
-        if let ColumnarValue::Array(array) = res {
-            assert_eq!(array.len(), 4);
-            let mut builder = Date32Array::builder(4);
-            builder.append_value(18_638);
-            builder.append_value(19_035);
-            builder.append_value(19_429);
-            builder.append_value(19_827);
-            assert_eq!(&builder.finish() as &dyn Array, array.as_ref());
-        } else {
-            panic!("Expected a columnar array")
-        }
-
-        //
-        // Fallible test cases
-        //
-
-        // invalid number of arguments
-        let res = make_date(&[ColumnarValue::Scalar(ScalarValue::Int32(Some(1)))]);
-        assert_eq!(
-            res.err().unwrap().strip_backtrace(),
-            "Execution error: make_date function requires 3 arguments, got 1"
-        );
-
-        // invalid type
-        let res = make_date(&[
-            ColumnarValue::Scalar(ScalarValue::IntervalYearMonth(Some(1))),
-            ColumnarValue::Scalar(ScalarValue::TimestampNanosecond(Some(1), None)),
-            ColumnarValue::Scalar(ScalarValue::TimestampNanosecond(Some(1), None)),
-        ]);
-        assert_eq!(
-            res.err().unwrap().strip_backtrace(),
-            "Arrow error: Cast error: Casting from Interval(YearMonth) to Int32 not supported"
-        );
-
-        // overflow of month
-        let res = make_date(&[
-            ColumnarValue::Scalar(ScalarValue::Int32(Some(2023))),
-            ColumnarValue::Scalar(ScalarValue::UInt64(Some(u64::MAX))),
-            ColumnarValue::Scalar(ScalarValue::Int32(Some(22))),
-        ]);
-        assert_eq!(
-            res.err().unwrap().strip_backtrace(),
-            "Arrow error: Cast error: Can't cast value 18446744073709551615 to type Int32"
-        );
-
-        // overflow of day
-        let res = make_date(&[
-            ColumnarValue::Scalar(ScalarValue::Int32(Some(2023))),
-            ColumnarValue::Scalar(ScalarValue::Int32(Some(22))),
-            ColumnarValue::Scalar(ScalarValue::UInt32(Some(u32::MAX))),
-        ]);
-        assert_eq!(
-            res.err().unwrap().strip_backtrace(),
-            "Arrow error: Cast error: Can't cast value 4294967295 to type Int32"
-        );
-    }
+    use datafusion_expr::{ColumnarValue, ScalarUDFImpl};
+    use std::sync::Arc;
 
     #[test]
     fn test_to_char() {
@@ -551,9 +323,9 @@ mod tests {
         ];
 
         for (value, format, expected) in scalar_data {
-            let result =
-                to_char(&[ColumnarValue::Scalar(value), ColumnarValue::Scalar(format)])
-                    .expect("that to_char parsed values without error");
+            let result = ToCharFunc::new()
+                .invoke(&[ColumnarValue::Scalar(value), ColumnarValue::Scalar(format)])
+                .expect("that to_char parsed values without error");
 
             if let ColumnarValue::Scalar(ScalarValue::Utf8(date)) = result {
                 assert_eq!(expected, date.unwrap());
@@ -625,11 +397,12 @@ mod tests {
         ];
 
         for (value, format, expected) in scalar_array_data {
-            let result = to_char(&[
-                ColumnarValue::Scalar(value),
-                ColumnarValue::Array(Arc::new(format) as ArrayRef),
-            ])
-            .expect("that to_char parsed values without error");
+            let result = ToCharFunc::new()
+                .invoke(&[
+                    ColumnarValue::Scalar(value),
+                    ColumnarValue::Array(Arc::new(format) as ArrayRef),
+                ])
+                .expect("that to_char parsed values without error");
 
             if let ColumnarValue::Scalar(ScalarValue::Utf8(date)) = result {
                 assert_eq!(expected, date.unwrap());
@@ -749,11 +522,12 @@ mod tests {
         ];
 
         for (value, format, expected) in array_scalar_data {
-            let result = to_char(&[
-                ColumnarValue::Array(value as ArrayRef),
-                ColumnarValue::Scalar(format),
-            ])
-            .expect("that to_char parsed values without error");
+            let result = ToCharFunc::new()
+                .invoke(&[
+                    ColumnarValue::Array(value as ArrayRef),
+                    ColumnarValue::Scalar(format),
+                ])
+                .expect("that to_char parsed values without error");
 
             if let ColumnarValue::Array(result) = result {
                 assert_eq!(result.len(), 2);
@@ -764,11 +538,12 @@ mod tests {
         }
 
         for (value, format, expected) in array_array_data {
-            let result = to_char(&[
-                ColumnarValue::Array(value),
-                ColumnarValue::Array(Arc::new(format) as ArrayRef),
-            ])
-            .expect("that to_char parsed values without error");
+            let result = ToCharFunc::new()
+                .invoke(&[
+                    ColumnarValue::Array(value),
+                    ColumnarValue::Array(Arc::new(format) as ArrayRef),
+                ])
+                .expect("that to_char parsed values without error");
 
             if let ColumnarValue::Array(result) = result {
                 assert_eq!(result.len(), 2);
@@ -783,14 +558,15 @@ mod tests {
         //
 
         // invalid number of arguments
-        let result = to_char(&[ColumnarValue::Scalar(ScalarValue::Int32(Some(1)))]);
+        let result = ToCharFunc::new()
+            .invoke(&[ColumnarValue::Scalar(ScalarValue::Int32(Some(1)))]);
         assert_eq!(
             result.err().unwrap().strip_backtrace(),
             "Execution error: to_char function requires 2 arguments, got 1"
         );
 
         // invalid type
-        let result = to_char(&[
+        let result = ToCharFunc::new().invoke(&[
             ColumnarValue::Scalar(ScalarValue::Int32(Some(1))),
             ColumnarValue::Scalar(ScalarValue::TimestampNanosecond(Some(1), None)),
         ]);
